Refactor: Post 폴더 따로 빼두고 파일 명을 index.tsx로 통일합니다
