+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,12 +1,15 @@
 ## 개요
+
 내용을 요약해주세요.
 
 ## 미리보기
+
 미리보기 사진이 있다면 올려주세요.
 
 ## 상세 설명
+
 자세한 설명을 적어주세요.
 
 ## 기타
 
-Close #{이슈_번호}
+Close #{이슈\_번호}
