+++ src/pages/Agreement/index.tsx
