+++ src/pages/Post/CommentList/index.tsx
@@ -1,16 +1,5 @@
 import { Stack } from '@chakra-ui/react'
-import styled from '@emotion/styled'
-
-const CommentItem = styled.li`
-  border-top: 1px solid #ccc;
-  border-bottom: 1px solid #ccc;
-  padding: 15px 0;
-  margin-top: -1px;
-  text-align: left;
-`
-const CommentAnonymousName = styled.div`
-  font-weight: bold;
-`
+import { CommentItem, CommentAnonymousName } from './styles'
 
 function CommentList() {
   return (
