+++ src/pages/Post/PetitionContents/styles.ts
